[O2B-645] Include tags update in run update
