+++ lib/database/repositories/RunTagsRepository.js
@@ -60,7 +60,7 @@ class RunTagsRepository extends Repository {
      * @return {Promise<void>} - delete result
      */
     async removeById(runId) {
-        return await RunTags.sequelize.getQueryInterface().bulkDelete('run_tags', { run_id: runId });
+        return RunTags.sequelize.getQueryInterface().bulkDelete('run_tags', { run_id: runId });
     }
 
     /**
@@ -68,8 +68,8 @@ class RunTagsRepository extends Repository {
      * @param {Array} entities - Tags & runId to associate
      * @return {*} Result of updating
      */
-    async updateMany(entities) {
-        return await RunTags.bulkCreate(entities, { returning: true });
+    async insertMany(entities) {
+        return RunTags.bulkCreate(entities, { returning: true });
     }
 }
