+++ lib/domain/dtos/UpdateRunDto.js
@@ -29,6 +29,9 @@ const EorReasonDto = Joi.object({
 const BodyDto = Joi.object({
     runQuality: Joi.string().valid('good', 'bad', 'test').default('test'),
     eorReasons: Joi.array().min(0).items(EorReasonDto).optional(),
+    tags: Joi.array()
+        .optional()
+        .items(Joi.number()),
 });
 
 const QueryDto = Joi.object({
