+++ lib/domain/dtos/index.js
@@ -44,7 +44,6 @@ const UpdateEnvironmentDto = require('./UpdateEnvironmentDto');
 const UpdateFlpDto = require('./UpdateFlpDto');
 const UpdateLhcFillDto = require('./UpdateLhcFillDto');
 const UpdateRunByRunNumberDto = require('./UpdateRunByRunNumberDto');
-const UpdateRunTagsDto = require('./UpdateRunTagsDto');
 const UpdateRunDto = require('./UpdateRunDto');
 const UpdateTagDto = require('./UpdateTagDto');
 const GetTagByNameDto = require('./GetTagByNameDto.js');
@@ -84,7 +83,6 @@ module.exports = {
     UpdateFlpDto,
     UpdateLhcFillDto,
     UpdateRunByRunNumberDto,
-    UpdateRunTagsDto,
     UpdateRunDto,
     UpdateTagDto,
 };
