+++ lib/public/views/Runs/Details/editTagsPanel.js
@@ -21,17 +21,9 @@ import { remoteDataTagPicker } from '../../../components/pickers/tag/remoteDataT
  * @param {Object} model Pass the model to access the defined functions.
  * @return {vnode} virtual node with representation of the panel
  */
-const editTagsPanel = (model) => h('#tags-selection.flex-column.w-30', [
-    h(
-        '.w-100.flex-row',
-        { style: 'justify-content: flex-end;' },
-        h('button.btn.btn-success.w-25#update-tags', {
-            onclick: () => model.runs.updateRunTags(),
-            title: 'This action will overwrite existing tags',
-            disabled: model.runs.editionTagPickerModel.hasOnlyDefaultSelection(),
-        }, 'Update Tags'),
-    ),
+const editTagsPanel = (model) => h(
+    '#tags-selection.flex-column.w-30',
     remoteDataTagPicker(model.tags.getTags(), model.runs.editionTagPickerModel),
-]);
+);
 
 export default editTagsPanel;
