+++ lib/public/views/Runs/Runs.js
@@ -35,7 +35,7 @@ export default class RunModel extends Observable {
         // Sub-models
         this._listingTagsFilterModel = new TagFilterModel();
         this._listingTagsFilterModel.observe(() => this._applyFilters());
-        this._editionTagPickerModel = new TagPickerModel();
+        this.editionTagPickerModel = new TagPickerModel();
 
         // Content
         this._currentPageRuns = RemoteData.notAsked();
@@ -99,7 +99,7 @@ export default class RunModel extends Observable {
 
         if (result.data) {
             this.run = RemoteData.success(result.data);
-            this._editionTagPickerModel = new TagPickerModel(this.run.payload.tags);
+            this.editionTagPickerModel = new TagPickerModel(this.run.payload.tags);
         } else {
             this.run = RemoteData.failure(result.errors || [
                 {
@@ -159,6 +159,7 @@ export default class RunModel extends Observable {
 
         if (result.data) {
             this.run = RemoteData.success(result.data);
+            this.editionTagPickerModel = new TagPickerModel(this.run.payload.tags);
         } else {
             this.run = RemoteData.failure(result.errors || [
                 {
@@ -667,44 +668,6 @@ export default class RunModel extends Observable {
     }
 
     /**
-     * Update (overwrite) run tags
-     * @return {void}
-     */
-    async updateRunTags() {
-        const { id } = this.run.payload;
-        const tags = this.editionTagPickerModel.selected;
-
-        this.run = RemoteData.loading();
-        this.notify();
-
-        const options = {
-            method: 'POST',
-            headers: {
-                Accept: 'application/json',
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({
-                tags: tags.map((tag) => tag.id),
-            }),
-        };
-        const response = await fetchClient(`/api/runs/${id}`, options);
-        const result = await response.json();
-
-        if (result.data) {
-            this.run = RemoteData.success(result.data);
-            this._editionTagPickerModel = new TagPickerModel(this.run.payload.tags);
-        } else {
-            RemoteData.failure(result.errors || [
-                {
-                    title: result.error,
-                    detail: result.message,
-                },
-            ]);
-        }
-        this.notify();
-    }
-
-    /**
      * Returns the state of table infinite scroll mode
      * @return {boolean} The state of table infinite scroll mode
      */
@@ -1168,7 +1131,7 @@ export default class RunModel extends Observable {
     /**
      * Method to update the property of a run;
      * If there is a missing key and value, all changes will be dropped
-     * @param {JSON} object containing a key and a value which represent the property of the run that is being changed
+     * @param {Object} object containing a key and a value which represent the property of the run that is being changed
      */
     set runChanges({ key, value }) {
         if (!key && !value) {
@@ -1237,6 +1200,18 @@ export default class RunModel extends Observable {
      */
     get editionTagPickerModel() {
         return this._editionTagPickerModel;
+    }
+
+    /**
+     * Set the current edition tag picker model and register required observers
+     *
+     * @param {TagPickerModel} editionTagPickerModel the new tag picker model
+     */
+    set editionTagPickerModel(editionTagPickerModel) {
+        this._editionTagPickerModel = editionTagPickerModel;
+        this._editionTagPickerModel.observe(() => {
+            this.runChanges = { key: 'tags', value: this.editionTagPickerModel.selected.map((tag) => tag.id) };
+        });
     }
 
     /**
