+++ lib/server/controllers/runs.controller.js
@@ -20,7 +20,6 @@ const {
         GetLogsByRunUseCase,
         GetRunUseCase,
         StartRunUseCase,
-        UpdateRunTagsUseCase,
         UpdateRunUseCase,
     },
 } = require('../../usecases');
@@ -30,7 +29,6 @@ const {
         GetAllRunsDto,
         GetRunDto,
         UpdateRunByRunNumberDto,
-        UpdateRunTagsDto,
         StartRunDto,
         UpdateRunDto,
     },
@@ -222,32 +220,6 @@ const listTagsByRunId = async (request, response, next) => {
 };
 
 /**
- * Update (overwrite) tags of specified run
- *
- * @param {Object} request The *request* object represents the HTTP request and has properties for the request query
- *                         string, parameters, body, HTTP headers, and so on.
- * @param {Object} response The *response* object represents the HTTP response that an Express app sends when it gets an
- *                          HTTP request.
- * @param {Function} next The *next* object represents the next middleware function which is used to pass control to the
- *                        next middleware function.
- * @returns {void}
- */
-const updateRunTags = async (request, response, next) => {
-    const value = await dtoValidator(UpdateRunTagsDto, request, response);
-    if (!value) {
-        return;
-    }
-
-    const { result, error } = await new UpdateRunTagsUseCase().execute(value);
-
-    if (error) {
-        response.status(Number(error.status)).json({ errors: [error] });
-    } else {
-        response.status(200).json({ data: result });
-    }
-};
-
-/**
  * Create a new Run
  *
  * @param {Object} request The *request* object represents the HTTP request and has properties for the request query
@@ -384,6 +356,5 @@ module.exports = {
     listTagsByRunId,
     startRun,
     updateRunByRunNumber,
-    updateRunTags,
     updateRun,
 };
