+++ lib/server/routers/runs.router.js
@@ -34,10 +34,6 @@ module.exports = {
                     controller: RunsController.getRunById,
                 },
                 {
-                    method: 'post',
-                    controller: RunsController.updateRunTags,
-                },
-                {
                     method: 'put',
                     controller: RunsController.updateRun,
                 },
