+++ lib/usecases/run/UpdateRunUseCase.js
@@ -16,6 +16,7 @@ const {
         RunRepository,
         EorReasonRepository,
         ReasonTypeRepository,
+        RunTagsRepository,
     },
     utilities: {
         TransactionHelper,
@@ -24,6 +25,7 @@ const {
 } = require('../../database');
 const { GetLhcFillUseCase } = require('../lhcFill');
 const GetRunUseCase = require('./GetRunUseCase');
+const GetAllTagsUseCase = require('../tag/GetAllTagsUseCase.js');
 
 /**
  * Update a run with provided values. For now we update only RunQuality
@@ -75,6 +77,7 @@ class UpdateRunUseCase {
                 fillNumber,
                 aliceL3Polarity,
                 aliceDipolePolarity,
+                tags,
             } = runUpdates;
 
             const queryBuilder = runId ? new QueryBuilder().where('id').is(runId) :
@@ -125,6 +128,24 @@ class UpdateRunUseCase {
                 }
                 if (eorReasons) {
                     await this._updateEorReasonsOnRun(runId, eorReasons);
+                }
+                if (tags) {
+                    const existingTags = await new GetAllTagsUseCase()
+                        .execute({ query: { page: { filter: { ids: tags } } } });
+                    const missingTags = tags.reduce((accumulator, tag) => {
+                        if (!existingTags.tags.find(({ id }) => id === tag)) {
+                            accumulator.push(tag);
+                        }
+                        return accumulator;
+                    }, []);
+
+                    if (missingTags.length > 0) {
+                        throw new Error(`Tags with these ids (${missingTags.join(', ')}) could not be found`);
+                    }
+
+                    // Remove existing tags of the run and insert new ones
+                    await RunTagsRepository.removeById(runId)
+                        .then(() => RunTagsRepository.insertMany(tags.map((tag) => ({ runId, tagId: tag }))));
                 }
                 await runObject.save();
                 return runObject;
