+++ lib/usecases/run/index.js
@@ -18,7 +18,6 @@ const GetFlpsByRunUseCase = require('./GetFlpsByRunUseCase');
 const GetLogsByRunUseCase = require('./GetLogsByRunUseCase');
 const GetRunUseCase = require('./GetRunUseCase');
 const StartRunUseCase = require('./StartRunUseCase');
-const UpdateRunTagsUseCase = require('../run/UpdateRunTagsUseCase');
 const UpdateRunUseCase = require('../run/UpdateRunUseCase');
 
 module.exports = {
@@ -29,6 +28,5 @@ module.exports = {
     GetLogsByRunUseCase,
     GetRunUseCase,
     StartRunUseCase,
-    UpdateRunTagsUseCase,
     UpdateRunUseCase,
 };
