+++ spec/openapi.yaml
@@ -1,4 +1,4 @@
-# Generated on Mon, 04 Jul 2022 16:01:42 GMT
+# Generated on Wed, 13 Jul 2022 06:24:30 GMT
 
 openapi: 3.0.0
 info:
@@ -877,12 +877,6 @@ components:
         application/json:
           schema:
             $ref: '#/components/schemas/UpdateRunQuality'
-    UpdateRunTags:
-      required: true
-      content:
-        application/json:
-          schema:
-            $ref: '#/components/schemas/UpdateRunTags'
   responses:
     ArrayOfAttachments:
       description: Expected response to a valid request.
@@ -2405,17 +2399,6 @@ components:
       properties:
         runQuality:
           $ref: '#/components/schemas/RunQuality'
-      additionalProperties: false
-    UpdateRunTags:
-      description: List of labels attached to something for the purpose of identification.
-      type: object
-      properties:
-        tags:
-          type: array
-          description: Array of tags IDs
-          items:
-            type: number
-            description: Tag ID
       additionalProperties: false
     User:
       description: Describes an intervention or an event that happened.
